Layout Impl of Scan API
