+++ vortex-file/src/file.rs
@@ -22,11 +22,11 @@ use vortex_array::expr::Expression;
 use vortex_array::expr::pruning::checked_pruning_expr;
 use vortex_error::VortexResult;
 use vortex_layout::LayoutReader;
+use vortex_layout::scan::layout::LayoutReaderDataSource;
+use vortex_layout::scan::scan_builder::ScanBuilder;
+use vortex_layout::scan::split_by::SplitBy;
 use vortex_layout::segments::SegmentSource;
-use vortex_scan::ScanBuilder;
-use vortex_scan::SplitBy;
-use vortex_scan::api::DataSourceRef;
-use vortex_scan::layout::LayoutReaderDataSource;
+use vortex_scan::DataSourceRef;
 use vortex_session::VortexSession;
 use vortex_utils::aliases::hash_map::HashMap;
 
@@ -90,7 +90,7 @@ impl VortexFile {
             .new_reader("".into(), segment_source, &self.session)
     }
 
-    /// Create a [`DataSource`](vortex_scan::api::DataSource) from this file for scanning.
+    /// Create a [`DataSource`](vortex_scan::DataSource) from this file for scanning.
     ///
     /// Wraps the file's layout reader with [`FileStatsLayoutReader`] (when file-level
     /// statistics are available) and [`LayoutReaderDataSource`].
