+++ vortex-file/src/tests.rs
@@ -62,8 +62,8 @@ use vortex_buffer::ByteBufferMut;
 use vortex_buffer::buffer;
 use vortex_error::VortexResult;
 use vortex_io::session::RuntimeSession;
+use vortex_layout::scan::scan_builder::ScanBuilder;
 use vortex_layout::session::LayoutSession;
-use vortex_scan::ScanBuilder;
 use vortex_session::VortexSession;
 
 use crate::OpenOptionsSessionExt;
