+++ vortex-layout/src/lib.rs
@@ -18,6 +18,7 @@ mod encoding;
 mod flatbuffers;
 mod layout;
 mod reader;
+pub mod scan;
 pub mod segments;
 pub mod sequence;
 pub mod session;
