+++ vortex-python/src/scan.rs
@@ -5,7 +5,7 @@ use pyo3::exceptions::PyIndexError;
 use pyo3::prelude::*;
 use vortex::array::ArrayRef;
 use vortex::array::DynArray;
-use vortex::scan::RepeatedScan;
+use vortex::layout::scan::repeated_scan::RepeatedScan;
 
 use crate::RUNTIME;
 use crate::error::PyVortexResult;
