+++ vortex-scan/Cargo.toml
@@ -17,28 +17,16 @@ version = { workspace = true }
 all-features = true
 
 [dependencies]
-arrow-array = { workspace = true }
-arrow-schema = { workspace = true }
 vortex-array = { workspace = true }
 vortex-buffer = { workspace = true }
 vortex-error = { workspace = true }
-vortex-io = { workspace = true }
-vortex-layout = { workspace = true }
 vortex-mask = { workspace = true }
-vortex-metrics = { workspace = true }
 vortex-session = { workspace = true }
 
 async-trait = { workspace = true }
-bit-vec = { workspace = true }
 futures = { workspace = true }
-itertools = { workspace = true }
-parking_lot = { workspace = true }
 roaring = { workspace = true }
-sketches-ddsketch = { workspace = true }
 tracing = { workspace = true }
-
-[dev-dependencies]
-vortex-layout = { workspace = true, features = ["_test-harness"] }
 
 [lints]
 workspace = true
